soc: qcom: ipa: change struct ipa_imm_cmd_hw_hdr_init_local
The bitfields defined in the ipa_imm_cmd_hw_hdr_init_local structure
add up to only 32 bits, while the "container" type is u64.  Define
this to use u32 as the container instead.  The layout of the
structure stays the same, because its size will be aligned to that
of its largest member (u64).

+++ drivers/net/ipa/ipahal_i.h
@@ -58,7 +58,7 @@ struct ipa_imm_cmd_hw_ip_fltrt_init {
  */
 struct ipa_imm_cmd_hw_hdr_init_local {
 	u64 hdr_table_addr;
-	u64 size_hdr_table	: 12,
+	u32 size_hdr_table	: 12,
 	    hdr_addr		: 16,
 	    rsvd		: 4;
 };
